Edit hxt_tetRefine.c: match types
Fixes build error on MacOS:
error: passing argument 3 of 'balanceRefineWork' from incompatible pointer type [-Wincompatible-pointer-types]

+++ contrib/hxt/tetMesh/src/hxt_tetRefine.c
@@ -364,7 +364,7 @@ HXTStatus hxtRefineTetrahedra(HXTMesh* mesh,
 {
   int maxThreads = omp_get_max_threads();
 
-  uint64_t* startTet; // see balanceRefineWork
+  size_t* startTet; // see balanceRefineWork
   HXT_CHECK( hxtMalloc(&startTet, (maxThreads+1)*sizeof(size_t)) );
 
   uint32_t* startPt; // see balanceRefineWork
